requirements.txt
openai~=1.35.3
click~=8.1.7
python-dotenv~=1.0.1
requests~=2.32.3